backend/requirements.txt
 
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.5
reportlab==4.0.7
pydantic==2.5.0
python-multipart==0.0.6
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1